Moved contracts folder under types
