+++ sources/packages/web/src/types/contracts/ConfigContract.ts
