+++ sources/packages/web/src/types/contracts/EducationContract.ts
